Added IGatewayProvider (see https://github.com/DV8FromTheWorld/JDA/issues/519)
