+++ src/main/java/net/dv8tion/jda/core/JDABuilder.java
@@ -24,6 +24,8 @@ import net.dv8tion.jda.core.exceptions.AccountTypeException;
 import net.dv8tion.jda.core.exceptions.RateLimitedException;
 import net.dv8tion.jda.core.hooks.IEventManager;
 import net.dv8tion.jda.core.managers.impl.PresenceImpl;
+import net.dv8tion.jda.core.requests.DefaultGatewayProvider;
+import net.dv8tion.jda.core.requests.IGatewayProvider;
 import net.dv8tion.jda.core.requests.SessionReconnectQueue;
 import net.dv8tion.jda.core.utils.Checks;
 import okhttp3.OkHttpClient;
@@ -46,6 +48,7 @@ public class JDABuilder
 {
     protected final List<Object> listeners;
 
+    protected IGatewayProvider gatewayProvider = null;
     protected SessionReconnectQueue reconnectQueue = null;
     protected ShardedRateLimiter shardRateLimiter = null;
     protected OkHttpClient.Builder httpClientBuilder = null;
@@ -81,6 +84,20 @@ public class JDABuilder
             throw new NullPointerException("Provided AccountType was null!");
         this.accountType = accountType;
         listeners = new LinkedList<>();
+    }
+
+    /**
+     * Sets the gateway provider that will be used to get the URL to connect the websocket to.
+     *
+     * @param  gatewayProvider
+     *         {@link net.dv8tion.jda.core.requests.IGatewayProvider IGatewayProvider} to use
+     *
+     * @return The {@link net.dv8tion.jda.core.JDABuilder JDABuilder} instance. Useful for chaining.
+     */
+    public JDABuilder setGatewayProvider(IGatewayProvider gatewayProvider)
+    {
+        this.gatewayProvider = gatewayProvider;
+        return this;
     }
 
     /**
@@ -538,7 +555,8 @@ public class JDABuilder
                 .setCacheGame(game)
                 .setCacheIdle(idle)
                 .setCacheStatus(status);
-        jda.login(token, shardInfo, reconnectQueue);
+        IGatewayProvider gatewayProvider = this.gatewayProvider == null ? new DefaultGatewayProvider(jda) : this.gatewayProvider;
+        jda.login(token, shardInfo, reconnectQueue, gatewayProvider);
         return jda;
     }
